requirements.txt
slack-bolt>=1.18.0
flask>=3.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
